Allow the 'updateStrategy' of the Daemonset to be configured in Helm
This PR enables the 'updateStrategy' of the Daemonset to be configured by Helm parameters.  This has two different parameters, one for Linux and one for Windows, following the existing pattern for these variables.  The default values are the same as the current defaults; RollingUpdate strategy with maxUnavailable set to 1, so behavior will change only if the user specifies a custom value.