feat: add keep-crd upgrade hook
The crds have been moved from the templates/ to crds/ folder. When helm
upgrade is run, helm will delete the crds because they're no longer in
the generated template. To prevent deletion, we patch the 2 CRDs with
the "helm.sh/resource-policy": "keep". Helm will skip deletion of
resources with these annotation. Also, converted the hooks to a job as
helm kills the pod before kubectl exec is run. With job, helm waits
until the pod is run and reaches completion.

Signed-off-by: Anish Ramasekar <anish.ramasekar@gmail.com>