fix: make manifest diff
A number of changes have been made directly to the manifest_staging
files - but those changes get overwritten whenever we use make manifest
to rebuild the files from the code annotations. This change adds a bunch
of sed mutations to the output of the code generator.

Updates to the tools (which determines which version of controller-gen
we use) have also not been applied.